lazy load helm wrapper in connector handler
