+++ pyproject.toml
@@ -37,6 +37,7 @@ pyhumps = "^3.7.3"
 cachetools = "^5.2.0"
 dictdiffer = "^0.9.0"
 python-schema-registry-client = "^2.4.1"
+sink = "^0.1.0b2"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.1.2"
